Say when polls start if they start later today
At the moment the election view just says that "polls are open till
10pm", which implies that polls are currently open. As polls in the UK
open at 7am and not midnight, this is vaguely misleading.

This commit attempts to add a "from 7am" if one is looking at the
election view on the day of the poll, but prior to polls opening. After
7am you don't care when they opened, just when it closes.