Re-add explicit time-zones to generated times
https://www.kanzaki.com/docs/ical/dtstart.html specifies that TZID can
be applied to individual times.

When we were (incorrectly) converting from UTC to Local, this was
getting automatically added, but now that we're not doing that
conversion,  apparently doesn't automatically add a timezone any
more (any any attempt I made to make it do so led to it incorrectly
offsetting times by an hour), so instead, manually set the time-zone up
on the event.

This can be verified by: