Restrict additional manifestos to specific parties
In #1955 I failed to consider
how not filtering on party would effect candidates within the list of
eligible ballot ids. This led to an additional manifesto from another party
to be shown on all candidates listed on these ballots. This fix adds a test
that mimics this bug and limits additional manifestos to a list of party ids.