handle None return value from get_election_timetable
I don't think this code is really right tbh,
but I'm doing it anyway

Fundamentally, I don't think get_election_timetable() should
really return None but sometimes it does and we're not
handling that in loads of places.

For now, I am choosing to draw a line at this level of
abstraction, deal with the fact that
get_election_timetable() sometimes returns None
as best we can here and say this needs fixing another day.

I don't think this is a huge problem for real-world data.
This is mainly to deal with the fact there are loads of tests
that don't correctly set up their test data so for example
there are loads of posts with no territory set under test.