Properly take territory_code from division or organisation
`ee_data.get("division")` can return `None`, so was failing for
ballots with no division.

This change means we always default to an empty dict, and fall back
to taking the territory_code from the organisation when there
is no division.

Tested on the current set of elections that has ballots with no
divisions.