refactor set_territory
territory_code is now a required field on divisions in EE so we don't need to account for it being empty with '-'. We still have elections with no divisions, though, so we still want to fall back to the org's territory_code in that case. And terrority_code on orgs is not a required field yet, so we still need to account for it being blank.